Move /acl module to src/
- Move Core\ACL::contactAutocomplete to Module\Search\Acl
- Rename module path to /search/acl
