+++ view/js/main.js
@@ -466,7 +466,7 @@ function updateConvItems(data) {
 		$('body').css('cursor', 'auto');
 	}
 	/* autocomplete @nicknames */
-	$(".comment-edit-form  textarea").editor_autocomplete(baseurl+"/acl");
+	$(".comment-edit-form  textarea").editor_autocomplete(baseurl + '/search/acl');
 	/* autocomplete bbcode */
 	$(".comment-edit-form  textarea").bbco_autocomplete('bbcode');
 }
