+++ view/templates/contacts-head.tpl
@@ -1,7 +1,7 @@
 
 <script>
 $(document).ready(function() {
-	$("#contacts-search").contact_autocomplete(baseurl + '/acl', 'a', true);
+	$("#contacts-search").contact_autocomplete(baseurl + '/search/acl', 'a', true);
 });
 </script>
