+++ view/templates/message-head.tpl
@@ -1,7 +1,7 @@
 
 <script>
 $(document).ready(function() {
-	$("#recip").name_autocomplete(baseurl + '/acl', 'm', false, function(data) {
+	$("#recip").name_autocomplete(baseurl + '/search/acl', 'm', false, function(data) {
 		$("#recip-complete").val(data.id);
 	});
 });
