+++ view/templates/poke_head.tpl
@@ -1,7 +1,7 @@
 
 <script>
 $(document).ready(function() {
-	$("#poke-recip").name_autocomplete(baseurl + '/acl', 'a', true, function(data) {
+	$("#poke-recip").name_autocomplete(baseurl + '/search/acl', 'a', true, function(data) {
 		$("#poke-recip-complete").val(data.id);
 	});
 });
