+++ view/theme/frio/templates/msg-header.tpl
@@ -1,6 +1,6 @@
 
 <script type="text/javascript">
-	$("#comment-edit-text-input").editor_autocomplete(baseurl+"/acl");
+	$("#comment-edit-text-input").editor_autocomplete(baseurl + '/search/acl');
 
 	$(document).ready(function() {
 		$("#comment-edit-text-input").bbco_autocomplete('bbcode');
