+++ view/theme/frio/templates/nav_head.tpl
@@ -1,6 +1,6 @@
 
 <script type="text/javascript">
 $(document).ready(function() {
-	$("#nav-search-input-field").search_autocomplete(baseurl + '/acl');	
+	$("#nav-search-input-field").search_autocomplete(baseurl + '/search/acl');
 });
 </script>
