+++ view/theme/vier/templates/nav_head.tpl
@@ -1,6 +1,6 @@
 
 <script>
 $(document).ready(function() {
-	$("#nav-search-text").search_autocomplete(baseurl + '/acl');
+	$("#nav-search-text").search_autocomplete(baseurl + '/search/acl');
 });
 </script>
